feat!: move derive macro from `uniplate::derive::Uniplate` to `uniplate::Uniplate`
Reexport the derive macro as `uniplate::Uniplate` instead of
`uniplate::derive::Uniplate`. As this has the same name of the trait,
this simplifies the imports needed to use Uniplate.

BREAKING CHANGE: replace references to `uniplate::derive::Uniplate` with
`uniplate::Uniplate`.
