+++ uniplate/tests/derive-pass/enums/allow-enum-fields-with-trailing-commas.rs
@@ -1,4 +1,4 @@
-use uniplate::derive::Uniplate;
+use uniplate::Uniplate;
 
 #[derive(Uniplate,PartialEq,Eq,Clone)]
 #[uniplate()]
