+++ uniplate/tests/derive-pass/enums/allow-enum-variant-with-no-fields.rs
@@ -1,4 +1,4 @@
-use uniplate::derive::Uniplate;
+use uniplate::Uniplate;
 #[derive(Clone, PartialEq, Eq, Uniplate)]
 #[uniplate()]
 enum A {
