+++ uniplate/tests/derive-pass/structs/allow-tuple-struct.rs
@@ -1,4 +1,4 @@
-use uniplate::derive::Uniplate;
+use uniplate::Uniplate;
 
 #[derive(Uniplate, PartialEq, Eq, Clone)]
 #[uniplate()]
