+++ uniplate/tests/derive-pass/structs/issue-16-structs.rs
@@ -1,5 +1,5 @@
 use std::collections::VecDeque;
-use uniplate::{derive::Uniplate, Biplate};
+use uniplate::{Uniplate, Biplate};
 
 #[derive(Eq, PartialEq, Clone, Debug, Uniplate)]
 #[uniplate()]
