+++ uniplate/tests/derive-pass/structs/recursive-structs.rs
@@ -1,4 +1,3 @@
-use uniplate::derive::Uniplate;
 use uniplate::Uniplate;
 
 #[derive(Uniplate, PartialEq, Eq, Clone)]
